% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_map_spdf_maker.R
\name{crop_map_spdf_maker}
\alias{crop_map_spdf_maker}
\title{gg_rasta}
\usage{
crop_map_spdf_maker(input, crs = "auto", zoom_out = 1)
}
\arguments{
\item{input}{df or raster}

\item{crs}{projection of data provided ('longlat'/'cea'/'auto')}
}
\description{
A function to plot biodiversity data.
}
\examples{
crop_map_world(df)
crop_map_world(df,crs='longlat')
}
