% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_map_world.R
\name{crop_map_world}
\alias{crop_map_world}
\title{crop_map_world}
\usage{
crop_map_world(input, crs = "auto", zoom_out = 1)
}
\arguments{
\item{input}{df or raster}

\item{crs}{projection of data provided ('longlat'/'cea'/'auto')}

\item{zoom_out}{zoom level for the map}
}
\description{
A function to plot biodiversity data.
}
\examples{
crop_map_world(df)
crop_map_world(df,crs='longlat')
}
