% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_guesser.R
\name{crs_guesser}
\alias{crs_guesser}
\title{crs_guesser}
\usage{
crs_guesser(input)
}
\arguments{
\item{input}{df/raster/spdf}
}
\description{
A function to guess crs of input.
}
\examples{
crs=crs_guesser(input)
}
