% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_organizer_function.R
\name{df_organizer}
\alias{df_organizer}
\title{df_organizer}
\usage{
df_organizer(input)
}
\arguments{
\item{input}{df}
}
\description{
A function to organize df into value, y, x form.
}
\examples{
df=df_organizer(input)
}
