% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bar.R
\name{gg_bar}
\alias{gg_bar}
\title{gg_bar}
\usage{
gg_bar(input, x = NULL, y = NULL, fill = input[, 1])
}
\arguments{
\item{input}{df/list/raster}

\item{crs}{}

\item{zoom_out}{}
}
\description{
A function to plot bar charts.
}
\examples{
gg_bar(df)
gg_bar(df)
}
