% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bee.R
\name{gg_bee}
\alias{gg_bee}
\title{gg_bee}
\usage{
gg_bee(input, density = NULL, icon_color = "black",
  legend_title = "input", filling = NULL, na = "white", map = NULL,
  img = NULL, y_sf = NULL, x_sf = NULL, breaks = NULL,
  size = NULL, scal_cont = NULL, break_n = 7,
  low = "lightgoldenrod", high = "darkgoldenrod", alpha = 0.7,
  map_col = "grey30", panel_background = "grey95", width = 0.25,
  breaks_x = NULL, breaks_y = NULL, map_title = NULL,
  proj_b = NULL, cellsize = 2, break_hex = NULL, hex_col = "black",
  hex_width = 0.5, square = NULL)
}
\arguments{
\item{icon_color}{choose color to change icon}

\item{filling}{choose pallete to color the diferent levels}

\item{na}{color for NA cells}

\item{map}{map to plot on top of input}

\item{img}{forg/lizard/bird/mammal}

\item{y_sf}{y position for icon}

\item{x_sf}{x position for icon}

\item{breaks}{choose breaks for color legend}

\item{size}{size of icon}

\item{scal_cont}{TRUE for gradient scale}

\item{low}{color for gradient scale}

\item{high}{color for gradient scale}

\item{alpha}{icon transparency 0-1}

\item{map_col}{color for map border}

\item{panel_background}{color for plot background}

\item{width}{width of map}

\item{breaks_x}{choose breaks for x axis}

\item{proj_b}{TRUE if using Berhman projection - default: FALSE}

\item{cellsize}{chose size of grids - default is 2}

\item{break_hex}{Decide if you want the grid to be cropped by countries borders, default - NULL}

\item{hex_col}{hexagon color defualt -'black'}

\item{hex_width}{hexagon border width, default - 0.5}

\item{square}{TRUE if square grid, FALSE produces an hexagon grid, default - FALSE}

\item{breaks_n}{number of breaks for color legend}

\item{breaks_y=choose}{breaks for y axis)}

\item{labels}{information to display in the legend}

\item{resolution}{grid resolution}

\item{crs}{projection of data provided ('longlat'/'cea')}

\item{zoom_out}{choose the best zoom to visualize your data}
}
\description{
A function to plot biodiversity data in grids.
}
\examples{
gg_rasta(input,icon_color='blue',legend_title="WEGE",filling=c(pal(7),palred(5)), na='red',map=africa_shape_sp,
img=img,y_sf=-20,x_sf=-10,size=10,breaks=c(0,1,2,3,4))

gg_rasta(input)
}
