% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_picky.R
\name{gg_picky}
\alias{gg_picky}
\title{gg_picky}
\usage{
gg_picky(input, pal = "qual", show_c_names = NULL, gfun = "richness",
  icon_color = "black", legend_title = NULL, filling = NULL,
  na = "white", map = NULL, img = NULL, y_sf = NULL, x_sf = NULL,
  breaks = NULL, size = NULL, scal_cont = NULL, break_n = 7,
  low = "yellow", high = "blueviolet", alpha = 0.7,
  map_col = "grey30", panel_background = "grey95", width = 0.25,
  breaks_x = NULL, breaks_y = NULL, map_title = NULL, label = NULL,
  resolution = 1, crs = crs, exclude_0 = "auto",
  zoom_out = zoom_out, round_to = 3)
}
\arguments{
\item{icon_color}{choose color to change icon}

\item{filling}{choose pallete to color the diferent levels}

\item{na}{color for NA cells}

\item{map}{map to plot on top of input}

\item{img}{forg/lizard/bird/mammal}

\item{y_sf}{y position for icon}

\item{x_sf}{x position for icon}

\item{breaks}{choose breaks for color legend}

\item{size}{size of icon}

\item{scal_cont}{TRUE for gradient scale}

\item{low}{color for gradient scale}

\item{high}{color for gradient scale}

\item{alpha}{icon transparency 0-1}

\item{map_col}{color for map border}

\item{panel_background}{color for plot background}

\item{width}{width of map}

\item{breaks_x}{choose breaks for x axis}

\item{label}{information to display in the legend}

\item{resolution}{grid resolution}

\item{crs}{projection of data provided ('longlat'/'cea')}

\item{zoom_out}{choose the best zoom to visualize your data. Higher the number - the more perspective you get.}

\item{breaks_n}{number of breaks for color legend}

\item{breaks_y=choose}{breaks for y axis)}

\item{crs}{projection of data provided ('longlat'/'cea')}
}
\description{
A function to plot your map in 131 different color palettes.
}
\examples{
gg_picky(input,icon_color='blue',legend_title="WEGE",filling=c(pal(7),palred(5)), na='red',map=africa_shape_sp,
img=img,y_sf=-20,x_sf=-10,size=10,breaks=c(0,1,2,3,4))

gg_picky(input, pal='qual')
}
