% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_rasta.R
\name{gg_rasta}
\alias{gg_rasta}
\title{gg_rasta}
\usage{
gg_rasta(input, gfun = "richness", icon_color = "black",
  legend_title = "input", filling = NULL, na = "white", map = NULL,
  img = "mammal", y_sf = NULL, x_sf = NULL, breaks = NULL,
  size = NULL, scal_cont = NULL, break_n = 7, low = "khaki1",
  high = "blueviolet", alpha = 0.7, map_col = "grey30",
  panel_background = "grey95", width = 0.25, breaks_x = NULL,
  breaks_y = NULL, map_title = NULL, label = NULL, resolution = 1,
  zoom_out = 1, exclude_0 = "auto", round_to = 3, map_bg = NA,
  invert_palette = NULL, return = "map", rem_map_leg = NULL,
  rem_hist_leg = NULL, add_values_hist = NULL, silent = NULL)
}
\arguments{
\item{gfun}{= 'richness'/'sum'/'density'}

\item{icon_color}{choose color to change icon}

\item{filling}{choose pallete to color the diferent levels}

\item{na}{color for NA cells}

\item{map}{map to plot on top of input}

\item{img}{forg/lizard/bird/mammal}

\item{y_sf}{y position for icon}

\item{x_sf}{x position for icon}

\item{breaks}{choose breaks for color legend}

\item{size}{size of icon}

\item{scal_cont}{TRUE for gradient scale}

\item{low}{color for gradient scale}

\item{high}{color for gradient scale}

\item{alpha}{icon transparency 0-1}

\item{map_col}{color for map border}

\item{panel_background}{color for plot background}

\item{width}{width of map}

\item{breaks_x}{choose breaks for x axis}

\item{label}{information to display in the legend}

\item{resolution}{grid resolution}

\item{zoom_out}{choose the best zoom to visualize your data. Higher the number - the more perspective you get.}

\item{breaks_n}{number of breaks for color legend}

\item{breaks_y=choose}{breaks for y axis)}

\item{crs}{projection of data provided ('longlat'/'cea')}
}
\description{
A function to plot biodiversity data.
}
\examples{
gg_rasta(input,icon_color='blue',legend_title="WEGE",filling=c(pal(7),palred(5)), na='red',map=africa_shape_sp,
img=img,y_sf=-20,x_sf=-10,size=10,breaks=c(0,1,2,3,4))

gg_rasta(input)
}
